Small mistake with devices

+++ src/translation/translators.py
@@ -67,7 +67,7 @@ class TranslationInferenceBeamSearch:
         results = []
         for seq_ in src:
             seq = seq_.unsqueeze(0)
-            logits = self.model(seq, torch.tensor([[self.bos_token]]))
+            logits = self.model(seq, torch.tensor([[self.bos_token]]).type_as(src))
             initial_pred_token = torch.topk(logits[0], k=self.beam_size, dim=-1)
             generated_tokens = torch.cat((torch.full((self.beam_size, 1),
                                                      self.bos_token),
